Update code bits
